Modified signitature sheet for Ph.D. disseration
I modified the WSU.sty and WSUThesisTemplate.tex files so that they have the "Ph.D in Engineering Program" line as well. I broke the master and phd dissertation signature page in the style file into two separate commands so that it can be called from the main LaTeX file.